+++ seppuku/resources/codesnippets/printVector.c
@@ -1,5 +1,5 @@
 printf("\n");
-for(int i = 0; i < §ID§.rows; i++){
+for(size_t i = 0; i < §ID§.rows; i++){
     printf("%d %c", ((int*)§ID§.dataStart)[i], '\n');
 }
 printf("\n");
